pyproject.toml
[project]
name = "cloud-cost-kb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "gitpython>=3.1.46",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "neo4j>=6.1.0",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
    "xlrd>=2.0.2",
]
